Rename imageUploadPanel.js to uploadPanel.js
