+++ opentreemap/treemap/js/src/lib/uploadPanel.js
