+++ opentreemap/treemap/js/src/user.js
@@ -2,7 +2,7 @@
 
 var inlineEditForm = require('treemap/lib/inlineEditForm.js'),
     recentEdits = require('treemap/lib/recentEdits.js'),
-    imageUploadPanel = require('treemap/lib/imageUploadPanel.js'),
+    uploadPanel = require('treemap/lib/uploadPanel.js'),
     csrf = require('treemap/lib/csrf.js'),
     $ = require('jquery');
 
@@ -26,7 +26,7 @@ recentEdits.init({
     nextLink: '#recent-user-edits-next',
 });
 
-imageUploadPanel.init({
+uploadPanel.init({
     panelId: '#set-photo-modal',
     imageElement: '#user-photo'
 });
