+++ opentreemap/treemap/templates/treemap/map_feature_detail.html
@@ -27,7 +27,7 @@
     {% include "treemap/partials/upload_file.html" with panel_id="add-photo-modal" title="Add a Photo" %}
     {% include "treemap/partials/photo_social_media_sharing.html" %}
 
-    {# Modal for viewing and rotating photos, used by imageUploadPanel.js #}
+    {# Modal for viewing and rotating photos, used by uploadPanel.js #}
     <div id="photo-lightbox" class="modal fade" tabindex="-1" role="dialog" aria-hidden="true">
         <div class="modal-dialog">
             <div class="modal-content lightbox">
